requirements.txt
albumentations==2.0.5
h5py==3.13.0
hydra-core==1.3.2
nibabel==5.3.2
timm==1.0.15
tqdm==4.67.1
pandas==2.2.3
segmentation_models_pytorch==0.4.0
scikit-learn==1.6.1
torch==2.5.1+cu121
torchvision==0.20.1+cu121
